requirements.txt
fastapi
uvicorn
motor
pymongo
streamlit
requests
python-dotenv
